renamed `drivers` to `ext_drivers`
